### 2022-12-16 #67 @Robaina
Pynteny: a Python package to perform synteny-aware, profile HMM-based searches in sequence databases
Submitting Author: Semidán Robaina (@Robaina)
All current maintainers: @Robaina
Package Name: Pynteny
One-Line Description of Package: Query sequence database by HMMs arranged in predefined synteny structure
Repository Link:  https://github.com/Robaina/Pynteny
Version submitted:   v0.0.5
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7696204.svg)](https://doi.org/10.5281/zenodo.7696204)
Version accepted: 1.0.0
Date accepted (month-day-year): 03-10-2023

---

- Include a brief paragraph describing what your package does:j

Pynteny is Python tool to search for [synteny](https://en.wikipedia.org/wiki/Synteny) blocks in (prokaryotic) sequence data through [HMMs](https://www.bioinformatics.org/wiki/Hidden_Markov_Model) of the ORFs of interest and [HMMER](http://hmmer.janelia.org/). By leveraging genomic context information, Pynteny can be employed to decrease the uncertainty of functional annotation of unlabelled sequence data due to the effect of paralogs. Pynteny can be accessed (i) through the command line, (ii) as a Python module, or (iii) as a (locally served) web application.

> *Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [notes on categories][NotesOnCategories] of our guidebook.*

Pynteny's main objective is to provide a means to query NGS (unannotated) sequence databases, such as metagenomic/metatranscriptomic datasets using syntenic blocks (i.e. spatial arrangements of genes) rather than single target genes/protein domains. In this sense, I would classify Pynteny within Data Extraction.

- Who is the target audience, and what are the scientific applications of this package?

Pynteny was designed to be used by researchers working with large, unannotated sequence databases, such as those typically encountered in metagenomic analyses. It can be accessed through a command line interface or easily integrated into pipelines as a Python package. Pynteny can also be used through a graphical interface running locally in the browser, which is more suitable for educational purposes.

To the extent of my knowledge, there isn't any Python package that provides the functionality provided by Pynteny.

  - If you made a pre-submission inquiry, please paste the link to the corresponding issue, forum post, or other discussions or `@tag` the editor you contacted: https://github.com/pyOpenSci/software-review/issues/65

- [x] has an [OSI approved license][OsiApprovedLicense].
- [x] has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others: GitHub Action

I had submitted this package for publication at JOSS prior to pyOpenSci. The submission is currently under consideration for scope in this issue: https://github.com/openjournals/joss-reviews/issues/4978

This option will allow reviewers to open smaller issues that can then be linked to PR's rather than submitting a more dense text-based review. It will also allow you to demonstrate addressing the issue via PR links.

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package should it be accepted.

- [x] [Last but not least, please fill out our pre-review survey](https://forms.gle/F9mou7S3jhe8DMJ16). This helps us track
submissions and improve our peer review process. We will also ask our reviewers

[Editor and review templates can be found here]: https://www.pyopensci.org/software-peer-review/how-to/editors-guide.html and https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html

[NotesOnCategories]: https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/aims-and-scope.html?highlight=data#notes-on-categories

[Templates]: https://www.pyopensci.org/contributing-guide/appendices/templates.html
