### 2022-12-19 #67 @NickleDave
Thank you @Robaina for submitting Pynteny to pyOpenSci. As you noted, we have already discussed on #65 and you have done extensive development to prepare for submission.

Here are the initial editor checks:

  - As discussed in #65 scope here includes data extraction and I think also data retrieval, and the package can and has been used for educational purposes. I additionally checked "retrieval" and "education" above
- [ ] **Documentation** The package has sufficient documentation available online (README, sphinx docs) to allow us to evaluate package function and scope *without installing the package*. This includes:
  - [x] tutorials or vignettes that help a user understand how to use the package and what it can do for them (often these have a name like "Getting started")
    - see e.g. https://robaina.github.io/Pynteny/examples/example_cli/
  - [x] API documentation - this includes clearly written doc strings with variables defined using a standard docstring format
  - [x] README that clearly articulates the function of the package
  - [ ] Contributing documentation that details how to install a development environment and how to contribute to the package
    - there is not a separate contributing page yet but there is a section in the README: [environment ](https://github.com/Robaina/Pynteny#setting-up-a-developer-environment)
- [x] **Issue Submission Documentation**: All of the information is filled out in the `YAML` header of the issue (located at the top of the issue template).
  - Will need to update version, editor, reviewers
- [x] **Automated tests:** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.
- [x] **License:** The package has an [OSI approved license](https://opensource.org/licenses)
  - Uses Apache license
- [x] **README:** The package has a README with clear explanation of what the package does and instructions on how to install it along with development instructions.
- [ ] **Contributing statement:** The package has a contributing.md file that details how to contribute to the package.
  - Not yet, see above
- [x] **Package overlap:** That package doesn't fully overlap with functionality of other packages that have already been submitted to pyOpenSci

---
---

The author has done a ton of work to get ready for submission. Pynteny is looking really good. We will need a contributing.md before the review finishes but that's the only thing I really see missing right now.

Thank you again @Robaina.
We will get an editor and reviewers as soon as possible.
I need to confirm timelines with @lwasser but I would expect it might take a bit longer because of the holidays, my guess is two weeks.
Will let you know for sure by the end of this week, but wanted to make sure you knew we have seen your submission and we are beginning the review.
