### 2023-01-09 b2e2701
removed get example data
