### 2023-01-12 #67 @arianesasso
Dear Robaina,

I hope you had a great start to the year as well! I am sorry for the delay. We are still looking for our second reviewer.

However, I would like to introduce you to @Batalex, who is already on the job as the first reviewer 😄 . We have previously talked about the package, and he already has access to the [reviewer's guide](https://www.pyopensci.org/peer-review-guide/software-peer-review-guide/reviewer-guide.html), the reviewer template, and the Python packaging standards detailed in the [packaging guide](https://www.pyopensci.org/python-package-guide). So, we will be able to continue moving 🚀 .

And please feel free to ping me at any time if you need help or have questions!
