### 2023-01-13 855164c
rm ipynb from ci
