+++ .github/workflows/tests.yml
@@ -37,6 +37,7 @@ jobs:
       - name: create environment with mamba
         uses: conda-incubator/setup-miniconda@v2
         with:
+          python-version: 3.8
           mamba-version: "*"
           channels: conda-forge,bioconda,defaults
           auto-activate-base: false
