+++ envs/pynteny-dev.yml
@@ -5,7 +5,7 @@ channels:
   - semidan.robaina
   - conda-forge
 dependencies:
-  - python >= 3.10
+  - python >= 3.8
   - poetry >= 1.3
   - coverage
   - pip
