+++ .github/workflows/tests.yml
@@ -38,7 +38,7 @@ jobs:
         uses: conda-incubator/setup-miniconda@v2
         with:
           python-version: 3.8
-          mamba-version: "*"
+          # mamba-version: "*"
           channels: conda-forge,bioconda,defaults
           auto-activate-base: false
           activate-environment: tests_pynteny
