### 2023-01-13 #67 @Batalex
Hey there! I am almost done with my review. No promises, but I think I can deliver this weekend.
Looking forward to discussing it with you! 🙌
