### 2023-01-15 #67 @Batalex
Here comes a big one! @Robaina @arianesasso

# Foreword

I want to thank both the author and the editors from pyOpenSci for the opportunity to tackle this package review. The following comments reflect my experience as I maintain private packages, and all remarks come with pynteny's best interest at heart.

Some comments are regrouped at a "general" level if they concern the project structure or several files. If I encountered a situation that I had already pointed out on another file, I did not comment again. However, I encourage taking into account the remarks from one file to the others.

Several comments I made are only valid for the submitted version (0.0.5) and were fixed in the meantime. Nonetheless, I will mark off what has been done later.

With that being said, let us get started!

I used the black cat🐈‍⬛ to highlight my personal thoughts in the review template below.

- [x]  As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

    🐈‍⬛  **The submitted version uses a personal conda channel in README, which is something I would not recommend. However, it uses bioconda in the main branch, so it is all good.**

    🐈‍⬛ **There are some improvements to be made regarding running the examples locally, see below.**

    🐈‍⬛ **In Pynteny 0.0.5, contributions guidelines are included in the README file. A more detailed CONTRIBUTING file has been added since then.**

   🐈‍⬛ **I noticed that two of the three links to the Pynteny repo use the HTTP scheme, and the third uses HTTPS.**

- [x]  Package has a [README.md](http://readme.md/) file in the root directory.

    🐈‍⬛ **Additional badges have been added since the 0.0.5 release.**

- [ ]  Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the [README.md](http://readme.md/) file.

### Usability

🐈‍⬛ **I ran the package in WSL since Apple silicon is not supported. The installation went smoothly, except for the really long resolution time. I am glad to see that the main branch encourages to use mamba.**

    🐈‍⬛  **As far as I can tell, it seems to do what it claims.**

    🐈‍⬛ **NA, no performance claim to be found in the documentation.**

    🐈‍⬛ **Tests covers 45% of pynteny. The coverage is heterogenous: the application and CLI parts of the package are less tester than the rest.**

    🐈‍⬛ **Unit tests did not run on CI as of 0.0.5. This has changed since then. The documentation website is built and deployed using CI, nice!**

        🐈‍⬛ **The code base does not fully follow PEP 8. See below for more information.**

### Final approval (post-review)

Estimated hours spent reviewing: ~8 hours

---

### General comments:

- The package is very well laid out and easy to navigate.
- I would encourage the author to follow the stylistic style guide defined by the [PEP8](https://peps.python.org/pep-0008/). The most immediate benefit this would provide to the code base is ***********consistency*********** with the rest of the Python ecosystem and a lower entry bar for potential contributors. Two easy fixes to apply in this code base are:
    - Using a code formater such as [black](https://black.readthedocs.io/en/stable/index.html).
    - Using `snake_case` to name functions and methods rather than `camelCase`
- Another suggestion would be to use a `.editorconfig` [file](https://editorconfig.org). There again, this is a simple way to enforce consistency in the code base by defining a few options to control the files’ format. This is especially true for Python projects since mixing tabs, and space crashes the program. There are other benefits as well, e.g., removing trailing spaces and adding new lines at the end of files.
- There are a few discrepancies in types, some of them could be fixed by converting `argparse.ArgumentParser` instances to `CommandArgs` before using them
- Empty parentheses in class definitions without inheritance are not necessary.
- “magic values” (such as lists of columns) could be defined at the module level rather than in the class/function definition, but I am nitpicking.
- My personal recommendation is to define the text encoding whenever a text file descriptor is opened. Even if this is not feasible at the moment, this is a tiny step toward Windows compatibility, and it removes a little bit of uncertainty.
- There are a lot of places where the author uses classes with only static methods and no attributes. This adds an unnecessary layer of complexity to the code base, where one could simply define functions in their dedicated file if keeping the namespace clean is important.
- With the author's permission, I would be glad to provide a simple, sensitive configuration to get started with a few quality code tools.

### CLI

- The defaults paths are quite inconvenient if I run the cli from the installed package rather than the source directory, the config file as well as the downloaded database ends up in `/home/<user>/miniconda3/envs/Pynteny/lib/python3.10/site-packages/`

- **I recommend adding "Installation" & "Getting Started" sections in the documentation**. The rationale is as follows: any part of the project may be a potential user's first contact. This includes the README file, the package description on Pypi (not applicable here) / a conda channel, and the documentation.
- **I would advise adding instructions on how to locally build the documentation** in [CONTRIBUTING.md](http://contributing.md/) in the section

```
3. Improving the documentation
```

- The following mention is incorrect in the example notebook:
> `To follow this example, you don't need to download _E. coli's_ genome, since it has been already downloaded during Pynteny's installation.`
The MG1655.gb file is stored in the `tests` folder, which is not included in the package. Therefore, when the package is installed using conda, it is not downloaded.

### pynteny/__init__.py

- `import *` are usually frowned upon in Python, but IMO this is ok in this case

### pynteny/api.py

- `importlib.metadata` is only available in the standard library from 3.8 and onwards. This is partially fixed in the `main` branch. For good measure, I would increase the minimum Python version in the `pyproject.toml` file as well, even if poetry is only used as a build backend.
- The author could use a simple type hint in the `Command` class to avoid a linter warning in the `update` method:

class Command():
    """Parent class for Pynteny command
    """
    _args: CommandArgs

```

- Empty `__init__` methods are not necessary, the smaller the code base the better!
- 🔥There are really nice uses of the package’s metadata in `_repr_html_`  and `cite`
- I am pleased to see that arguments with `None` default value are correctly types as `optional`
- There is not need to wrap `Path(__file__).parent` in another `Path()` in the `getTestDataPath` method, since it returns a `Path` object as well. This comment applies to utils.py as well.

### pynteny/cli.py

- `args` is not used in the `Pynteny.app` method, so there is no need to parse it.
- I believe `getHelpStr` could be simpler by writing to a `io.StringIO` instead of a temp file.
- f-strings without placeholder are considered code smells
- Overall, this is a really nice example of a cli with subcommands, congrats!
- I saw a few unused `required` and `pynteny` assignment

### pynteny/filter.py

- typo L40 `specified`
- As far as I can tell, the `hmm_order_dict` is not needed in `SyntenyPatternFilters` init method.
- The lambda expr + map is quite difficult to read. A simpler alternative could be:

strand_map = dict(neg=-1, pos=1) # could be set as a module-wide constant
self.strand_order_pattern = [strand_map.get(strand, 0) for strand in parsed_structure["strands"]]
```

- `contains_hmm_pattern`, `contains_distance_pattern`, and `contains_strand_pattern` could use Python’s idioms and return boolean values instead of integers, this is not an issue with their use in pandas rolling method. Boolean values can be aggregated numerically in pandas. This would simplify those methods. In `contains_strand_pattern`, there is a `True` equality comparison.
- Using `inplace=True` in pandas operation is not recommended and could be deprecated: [https://github.com/pandas-dev/pandas/issues/16529](https://github.com/pandas-dev/pandas/issues/16529).
- The author could use the `size` method instead of a filter with a lambda in `getAllHMMhits`, but this is a matter of preference.
- There again, preferences, but the loop in `filterHitsBySyntenyStructure` could be made on a pandas’ groupby object. This is not a prominent feature, but one may do the following:

for group_key, group_df in df.groupby(key):
    pass
```

- the form `len(contig_hits.hmm.unique())` could be simplified to `contig_hits.hmm.nunique()`
- Using `iterrows` is quite inefficient if one only need to loop over the index. The author may use `for i in matched_rows.index`
- Idiomatic Python does not usually use getter methods such as `getSyntenyHits`, a public attribute would work just fine.
- I have not profiled the code to see if there is a performance bottleneck there, but I would use `itertuples` instead of `iterrows` in `addHMMmetaInfoToHits`. `itertuples` is much faster if you need to iterate over data frames one row at a time
- I suggest using `isinstance` rather than type equality in `filterFASTAbySyntenyStructure`. `isinstance` is a more robust (and marginally faster) solution because it allows nominal subtyping: any subtype of `str` or `list` would pass the test conditions. Using `isinstance` is therefore considered more pythonic than type equality.

### pynteny/hmm.py

- Nice use of properties
- Since the author is using pathlib, they might use `/` or `Path.joinpath` instead of `os.path.join`
- The author may use the `usecols` argument for `pd.read_csv`
- Why is the`try-except` block needed in `getHMMnamesByGeneSymbol`? (+ bare except is considered a code smell). The code would work without this. The same goes for `getHMMgeneID`, with a little twist to return `None`, and for `getMetaInfoForHMM`

### pynteny/parser.py

- The two classes in this file only use static methods and do not hold any attribute. Is there a reason for not replacing all those methods with simple functions?
Edit: event `preprocessing.py` does `isLegitSequence = RecordSequence.isLegitPeptideSequence`
- There are too many statements in the `try-except` block in `LabelParser.parse`. I do not know what could be failing (a `split`, an int cast?), and the exception caught is too broad. Tip: I suggest extracting `meta.split("_")` in a variable since it appears in four lines in a row
- Typing tuples is quite unintuitive: `tuple[str]` means “a tuple of exactly one st.” This means that  `splitStrandFromLocus` is incorrectly typed

### pynteny/preprocessing.py

- In FASTA class, we could replace the getter/setter with a property, which is more “pythonic”. Moreover, we could also make `_input_file_str` a read-only property.
- I would prefer if the inner function `unique_records`were defined outside the `removeDuplicates` method. I understand that the purpose is to generate the records lazily, which could still work as intended.
- `number_prodigal_record_fields` could be a global constant
- There again with `fromGenBankData`, I think the three inner functions would be better off defined outside the method. Contrary to the previous time, they do not use `self` and are even easier to extract. This change would reduce `fromGenBankData` complexity and make it even easier to debug.
- If pynteny targets python ≥3.10, then it may use [https://docs.python.org/3/whatsnew/3.10.html#parenthesized-context-managers](https://docs.python.org/3/whatsnew/3.10.html#parenthesized-context-managers) in `GeneAnnotator`

### pynteny/subcommands.py

- There is a piece of code duplicated four times. The following could be extracted in its own function.

if args.logfile is None:
    args.logfile = Path(os.devnull)
elif not Path(args.logfile.parent).exists():
    Path(args.logfile.parent).mkdir(parents=True)
logging.basicConfig(format='%(asctime)s | %(levelname)s: %(message)s',
                    handlers=[
                        logging.FileHandler(args.logfile),
                        logging.StreamHandler(sys.stdout)
                        ],
                    level=logging.NOTSET)
logger = logging.getLogger(__name__)
```

In addition, this would limit the complexity of `synteny_search`

- `wget` is quite old and does not seem to be active. How about replacing it with a more maintained alternative? httpx, requests
- As I said in the general comments, I would prefer if the default dirs were not relative to the package files. When installing pynteny inside a virtual env, this means that the database could be downloaded inside the venv, in a totally different place from the current working dir. I suggest using `~/.pynteny`

### pynteny/utils.py

- As I said in the comments above, using the file location is quite inconvenient because when installed by conda it ends up in the venv
- Is there a reason for not using [https://docs.python.org/3/library/tarfile.html?highlight=tar#tarfile.is_tarfile](https://docs.python.org/3/library/tarfile.html?highlight=tar#tarfile.is_tarfile)?

### pynteny/wrappers.py

- ✅

### pynteny/app

- The embedded streamlit app is really cool 🔥
- There again, it seems a bit odd to have classes with only static methods
