### 2023-01-15 #67 @Robaina
Hey @Batalex ,

thanks so much for your very detailed review!  I quickly run through your comments and II see that they will be very useful. I'll address them carefully in the following days. But wanted to thank you now.
