### 2023-01-16 #67 @arianesasso
@Batalex, what a beautiful and thoughtful review! I really enjoyed the 🐈‍⬛ haha and, of course, all the meaningful recommendations you made! Thank you for all your effort 🙏🏼.

@Robaina, hopefully, we will have the second reviewer onboarded soon. I will keep you posted!
