### 2023-01-17 Robaina/Pynteny#39 @Robaina
Refactor code to meet PEP8
Following recommendations in https://github.com/pyOpenSci/software-review/issues/67#issuecomment-1384640221
