+++ pynteny/app/components.py
@@ -18,53 +18,74 @@ class Sidebar:
         st.sidebar.image(
             st.session_state.sidebar_icon,
             use_column_width=True,
-            caption=f"Synteny-aware HMM searches made easy (v{__version__})"
+            caption=f"Synteny-aware HMM searches made easy (v{__version__})",
         )
 
         st.sidebar.text(" ")
         st.sidebar.text(" ")
 
         with st.sidebar.expander("Current output directory:", expanded=True):
-            if (st.session_state.outdir is not None) and (st.session_state.outdir.exists()):
+            if (st.session_state.outdir is not None) and (
+                st.session_state.outdir.exists()
+            ):
                 files_div = st.empty()
                 with files_div.container():
-                    FileManager.show_files_in_dir(st.session_state.outdir, sidebar=False)
-            
+                    FileManager.show_files_in_dir(
+                        st.session_state.outdir, sidebar=False
+                    )
+
             st.text(" ")
-            st.button("Select directory", on_click=Callbacks.updateOutputDir)
-            col1, col2 = st.columns([.5, .5])
+            st.button("Select directory", on_click=Callbacks.update_output_dir)
+            col1, col2 = st.columns([0.5, 0.5])
             with col1:
-                st.text_input("",
-                            value="", max_chars=None,
-                            key="subdirectory", on_change=Callbacks.updateOutputSubdirectory,
-                            placeholder="Create subdirectory")
+                st.text_input(
+                    "",
+                    value="",
+                    max_chars=None,
+                    key="subdirectory",
+                    on_change=Callbacks.update_output_subdirectory,
+                    placeholder="Create subdirectory",
+                )
             with col2:
-                st.text_input("",
-                            value="", max_chars=None,
-                            key="prefix", on_change=Callbacks.updateOutputPrefix,
-                            placeholder="Enter output prefix")
+                st.text_input(
+                    "",
+                    value="",
+                    max_chars=None,
+                    key="prefix",
+                    on_change=Callbacks.update_output_prefix,
+                    placeholder="Enter output prefix",
+                )
 
         with st.sidebar.expander("Advanced parameters:", expanded=False):
 
             st.markdown("Select custom HMM database")
             col1, col2 = st.columns([1, 1])
             with col1:
-                st.button("HMM directory", on_click=Callbacks.selectHMMdir)
+                st.button("HMM directory", on_click=Callbacks.select_HMM_dir)
             with col2:
-                st.button("HMM metadata", on_click=Callbacks.selectHMMmeta)
-            
+                st.button("HMM metadata", on_click=Callbacks.select_HMM_meta)
+
             col1, col2, col3 = st.columns([0.4, 0.3, 0.3])
             with col1:
                 st.markdown(("Processes:"))
-                st.slider("Processes",
-                        min_value=1, max_value=os.cpu_count(),
-                        value=os.cpu_count() - 1, step=1,
-                        on_change=Callbacks.setNumberOfProcesses, key="processes"
-                        )
+                st.slider(
+                    "Processes",
+                    min_value=1,
+                    max_value=os.cpu_count(),
+                    value=os.cpu_count() - 1,
+                    step=1,
+                    on_change=Callbacks.set_number_of_processes,
+                    key="processes",
+                )
             with col2:
                 st.markdown(("Output log:"))
-                output_log = st.select_slider("", options=["No", "Yes"], value="Yes",
-                             on_change=Callbacks.selectLogPath, key="log")
+                output_log = st.select_slider(
+                    "",
+                    options=["No", "Yes"],
+                    value="Yes",
[…]
