+++ pynteny/utils.py
@@ -20,26 +20,25 @@ from multiprocessing import Pool
 logger = logging.getLogger(__name__)
 
 
-class CommandArgs():
-    """Base class to hold command line arguments.
-    """
+class CommandArgs:
+    """Base class to hold command line arguments."""
+
     def __init__(self, **kwargs):
         self.__dict__.update(kwargs)
 
-        
-class ConfigParser():
-    """Handle Pynteny configuration file.
-    """
+
+class ConfigParser:
+    """Handle Pynteny configuration file."""
+
     def __init__(self, config_file: Path) -> None:
         self._config_file = Path(config_file)
         self._config = self.get_config()
-    
+
     @classmethod
     def get_default_config(cls):
-        """Initialize ConfigParser with default config file.
-        """
+        """Initialize ConfigParser with default config file."""
         return cls(cls.initialize_config_file())
-    
+
     @staticmethod
     def initialize_config_file() -> Path:
         """Initialize empty config file.
@@ -55,15 +54,14 @@ class ConfigParser():
                 "data_downloaded": False,
                 "PGAP_database": "",
                 "PGAP_meta_file": "",
-                "streamlit_process": ""
+                "streamlit_process": "",
             }
-            with open(config_file, 'w') as f:
+            with open(config_file, "w") as f:
                 json.dump(config, f, indent=4)
         return config_file
-    
+
     def get_config_path(self) -> Path:
-        """Show config file path.
-        """
+        """Show config file path."""
         return self._config_file
 
     def get_config(self) -> dict:
@@ -72,14 +70,13 @@ class ConfigParser():
         Returns:
             dict: dict containing fields and values of config file.
         """
-        with open(self._config_file, 'r') as file:
+        with open(self._config_file, "r") as file:
             config = json.loads(file.read())
         return config
 
     def write_config(self) -> None:
-        """Write config dict to file.
-        """
-        with open(self._config_file, 'w') as f:
+        """Write config dict to file."""
+        with open(self._config_file, "w") as f:
             json.dump(self._config, f, indent=4)
 
     def update_config(self, key: str, value: str) -> None:
@@ -104,11 +101,14 @@ class ConfigParser():
         return self._config[key]
 
 
-def setDefaultOutputPath(input_path: Path, tag: str = None,
-                         extension: str = None,
-                         only_filename: bool = False,
-                         only_basename: bool = False,
-                         only_dirname: bool = False) -> Path:
+def set_default_output_path(
+    input_path: Path,
+    tag: str = None,
+    extension: str = None,
+    only_filename: bool = False,
+    only_basename: bool = False,
+    only_dirname: bool = False,
+) -> Path:
     """Utility function to generate a default path to output file
     or directory based on an input file name and path.
 
@@ -128,8 +128,8 @@ def setDefaultOutputPath(input_path: Path, tag: str = None,
     if extension is None:
         extension = ext
     if tag is None:
-        tag = ''
-    default_file = f'{fname}{tag}{extension}'
+        tag = ""
+    default_file = f"{fname}{tag}{extension}"
     if only_basename:
         return Path(fname)
     if only_filename:
@@ -139,10 +139,13 @@ def setDefaultOutputPath(input_path: Path, tag: str = None,
     else:
         return Path(os.path.abspath(os.path.join(dirname, default_file)))
 
-def terminalExecute(command_str: str,
-                    suppress_shell_output=False,
[…]

[…]
