+++ tests/test_integration_search.py
@@ -13,6 +13,7 @@ from pynteny.api import Search
 
 this_file_dir = Path(Path(__file__).parent)
 
+
 class TestSyntenySearch(unittest.TestCase):
     def test_search(self):
         with TemporaryDirectory() as tempdir:
@@ -28,19 +29,19 @@ class TestSyntenySearch(unittest.TestCase):
                 reuse=True,
                 logfile=None,
                 processes=None,
-                unordered=False
+                unordered=False,
             )
-            synhits = search.run().getSyntenyHits()
+            synhits = search.run().get_synteny_hits()
             config = Path(this_file_dir.parent) / "config.json"
             if config.exists():
                 config.unlink()
         hit_labels = [
-            'b0071__U00096_71_78847_79453_neg',
-            'b0072__U00096_72_79463_80864_neg',
-            'b0074__U00096_74_81957_83529_neg'
-            ]
+            "b0071__U00096_71_78847_79453_neg",
+            "b0072__U00096_72_79463_80864_neg",
+            "b0074__U00096_74_81957_83529_neg",
+        ]
         self.assertListEqual(
             synhits.full_label.values.tolist(),
             hit_labels,
-            "Failed finding synteny hits"
+            "Failed finding synteny hits",
         )
