### 2023-01-19 c1aae03
added tests, logging.shutdown()
