+++ .github/workflows/tests.yml
@@ -52,3 +52,5 @@ jobs:
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v3
+        with:
+          version: "v0.1.15"
