+++ pyproject.toml
@@ -10,8 +10,8 @@ license = "Apache-2.0"
 authors = ["Semidán Robaina Estévez <srobaina@ull.edu.es>"]
 maintainers = ["Semidán Robaina Estévez <srobaina@ull.edu.es>"]
 readme = "README.md"
-homepage = "http://github.com/robaina/Pynteny"
-repository = "http://github.com/robaina/Pynteny"
+homepage = "https://github.com/robaina/Pynteny"
+repository = "https://github.com/robaina/Pynteny"
 documentation = "https://robaina.github.io/Pynteny"
 keywords = ["synteny", "hmm", "metagenomics", "bioinformatics", "hmmer"]
 classifiers = [
