+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "pynteny"
-version = "0.0.5"
+version = "0.0.6"
 description = "Synteny-aware hmm searches made easy in Python"
 license = "Apache-2.0"
 authors = ["Semidán Robaina Estévez <srobaina@ull.edu.es>"]
@@ -25,6 +25,6 @@ include = [
     "pynteny/app/assets/img/*"
 ]
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.8"
 [tool.poetry.scripts]
 pynteny = "pynteny.cli:main"
