### 2023-01-24 3859072
getter as property
