### 2023-01-25 7b21f36
extract static classes from parse.py
