### 2023-01-25 12ee4ec
extrac nested in LabelledFASTA
