### 2023-01-25 79b0849
extrac nested in LabelledFASTA
