### 2023-01-25 932c8cc
extrac nested in LabelledFASTA
