### 2023-01-25 5d85afd
extract nested functions in LabelledFasta
