### 2023-01-25 8427d87
extract nested functions in LabelledFasta
