+++ pynteny/preprocessing.py
@@ -10,13 +10,14 @@ Tools to preprocess sequence databases
 """
 
 from __future__ import annotations
+from typing import TextIO
 import sys
 import os
 import logging
 import tempfile
 from pathlib import Path
 
-from Bio import SeqIO
+from Bio import SeqIO, SeqRecord, SeqFeature
 import pyfastx
 
 import pynteny.utils as utils
@@ -385,50 +386,20 @@ class LabelledFASTA(FASTA):
                 Path(gbk_files.pop().parent) / f"{prefix}sequence_database.fasta"
             )
 
-        def get_label_str(gbk_contig, feature):
-            name = feature.qualifiers["locus_tag"][0].replace("_", ".")
-            start, end, strand = (
-                str(feature.location.start),
-                str(feature.location.end),
-                feature.location.strand,
-            )
-            start = start.replace(">", "").replace("<", "")
-            end = end.replace(">", "").replace("<", "")
-            strand_sense = "neg" if strand == -1 else ("pos" if strand == 1 else "")
-            return f">{name}__{gbk_contig.name.replace('_', '')}_{gene_counter}_{start}_{end}_{strand_sense}\n"
-
-        if nucleotide:
-
-            def write_record(gbk_contig, feature, outfile, gene_counter):
-                header = get_label_str(gbk_contig, feature)
-                sequence = str(feature.extract(gbk_contig).seq)
-                outfile.write(header)
-                outfile.write(sequence + "\n")
-                gene_counter += 1
-                return gene_counter
-
-        else:
-
-            def write_record(gbk_contig, feature, outfile, gene_counter):
-                if "translation" in feature.qualifiers:
-                    header = get_label_str(gbk_contig, feature)
-                    sequence = feature.qualifiers["translation"][0]
-                    outfile.write(header)
-                    outfile.write(sequence + "\n")
-                    gene_counter += 1
-                return gene_counter
-
         with open(output_file, "w+") as outfile:
             for gbk_contig in gbk_contigs:
                 gene_counter = 0
                 for feature in gbk_contig.features:
                     if "cds" in feature.type.lower():
-                        gene_counter = write_record(
-                            gbk_contig, feature, outfile, gene_counter
+                        gene_counter = cls.write_record(
+                            gbk_contig, feature, outfile, gene_counter, nucleotide
                         )
         return cls(output_file)
 
-    def get_label_str(self, gbk_contig, feature, gene_counter):
+    @staticmethod
+    def get_label_str(
+        gbk_contig: SeqRecord, feature: SeqFeature, gene_counter: int
+    ) -> str:
         name = feature.qualifiers["locus_tag"][0].replace("_", ".")
         start, end, strand = (
             str(feature.location.start),
@@ -440,13 +411,24 @@ class LabelledFASTA(FASTA):
         strand_sense = "neg" if strand == -1 else ("pos" if strand == 1 else "")
         return f">{name}__{gbk_contig.name.replace('_', '')}_{gene_counter}_{start}_{end}_{strand_sense}\n"
 
-    def write_record(self, gbk_contig, feature, outfile, gene_counter):
-        if "translation" in feature.qualifiers:
-            header = self.get_label_str(gbk_contig, feature)
+    @staticmethod
+    def write_record(
+        gbk_contig: SeqRecord,
+        feature: SeqFeature,
+        outfile: TextIO,
+        gene_counter: int,
+        nucleotide: bool = False,
+    ) -> int:
+        header = LabelledFASTA.get_label_str(gbk_contig, feature, gene_counter)
+        if (not nucleotide) and ("translation" in feature.qualifiers):
             sequence = feature.qualifiers["translation"][0]
-            outfile.write(header)
[…]
