### 2023-01-25 9b29294
extract nested functions in LabelledFasta
