### 2023-01-25 88145a3
use only seqkit to remove duplicates
