### 2023-01-25 4e7504f
update logger init
