### 2023-01-25 3509fbd
extracted static classes from parser
