### 2023-01-25 bea7574
removed static classes
