+++ pynteny/filter.py
@@ -291,7 +291,6 @@ class SyntenyHMMfilter:
             }
 
             if contig_hits.hmm.nunique() >= self._n_hmm_groups:
-
                 hmm_match = contig_hits.hmm_code.rolling(
                     window=self._n_hmm_groups
                 ).apply(filters.contains_hmm_pattern)
