### 2023-01-26 04a5f15
add setup to docs, add building docs to contributing
