### 2023-01-26 ad995fc
add setup to docs, add building docs to contributing
