### 2023-01-26 f00e26f
add setup to docs, add building docs to contributing
