+++ pynteny/parser.py
@@ -1,289 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""
-Tools to parse synteny structure
-"""
-
-from __future__ import annotations
-import sys
-import logging
-from pathlib import Path
-
-import pandas as pd
-
-from pynteny.hmm import PGAP
-from pynteny.utils import is_right_list_nested_type
-
-logger = logging.getLogger(__name__)
-
-
-class LabelParser:
-    """Parse entry label to extract coded info"""
-
-    @staticmethod
-    def parse(label: str) -> dict:
-        """Parse sequence labels to obtain contig and locus info
-
-        Args:
-            label (str): sequence label
-
-        Returns:
-            dict: dictionary with parsed label info.
-        """
-        parsed_dict = {
-            "full_label": label,
-            "gene_id": "",
-            "contig": "",
-            "gene_pos": None,
-            "locus_pos": None,
-            "strand": "",
-        }
-
-        if label.count("__") > 1:
-            logger.error("Invalid format of record label string")
-            sys.exit(1)
-
-        entry = label.split("__")[0]
-        meta = label.split("__")[1]
-        meta_items = meta.split("_")
-
-        strand = meta_items[-1]
-        locus_pos = tuple([int(pos) for pos in meta_items[-3:-1]])
-        gene_pos = int(meta_items[-4])
-        contig = "_".join(meta_items[:-4])
-
-        parsed_dict["gene_id"] = entry
-        parsed_dict["contig"] = contig
-        parsed_dict["gene_pos"] = gene_pos
-        parsed_dict["locus_pos"] = locus_pos
-        parsed_dict["strand"] = strand
-
-        return parsed_dict
-
-    @staticmethod
-    def parse_from_list(labels: list[str]) -> pd.DataFrame:
-        """Parse labels in list of labels and return DataFrame.
-
-        Args:
-            labels (list, optional): list of labels as stringgs.
-
-        Returns:
-            pd.DataFrame: Dataframe containing parsed information from labels.
-        """
-        return pd.DataFrame([LabelParser.parse(label) for label in labels])
-
-
-class SyntenyParser:
-    """Tools to parse synteny structure strings"""
-
-    @staticmethod
-    def reformat_synteny_structure(synteny_structure: str) -> str:
-        """Remove illegal symbols and extra white space."""
-        synteny_structure = (
-            synteny_structure.replace(" |", "|").replace("| ", "|").strip()
-        )
-        return synteny_structure
-
-    @staticmethod
-    def contains_HMM_groups(synteny_structure: str) -> bool:
-        """Check whether structure contains groups of gene-equivalent HMMs."""
-        return "|" in synteny_structure
-
-    @staticmethod
-    def is_valid_structure(synteny_structure: str) -> bool:
-        """Validate synteny structure format."""
-        synteny_structure = (
-            synteny_structure.replace(" |", "|").replace("| ", "|").strip()
-        )
-        parsed_struc = SyntenyParser.parse_synteny_structure(synteny_structure)
-        right_type = all(
-            (
-                is_right_list_nested_type(parsed_struc["hmm_groups"], str),
-                is_right_list_nested_type(parsed_struc["strands"], str),
-                is_right_list_nested_type(parsed_struc["distances"], int),
-            )
-        )
-        right_format = len(parsed_struc["hmm_groups"]) == len(
-            parsed_struc["strands"]
-        ) and len(parsed_struc["hmm_groups"]) == (len(parsed_struc["distances"]) + 1)
-        return False if (not right_type or not right_format) else True
-
-    @staticmethod
-    def split_strand_from_locus(
-        locus_str: str, parsed_symbol: bool = True
-    ) -> tuple[str, ...]:
-        """Split strand info from locus tag / HMM model.
-
-        Args:
[…]
