+++ pynteny/utils.py
@@ -196,12 +196,7 @@ def is_tar_file(tar_file: Path) -> bool:
     Returns:
         bool: whether file is compressed or not.
     """
-    tar_file_str = tar_file.as_posix()
-    return (
-        (tar_file_str.endswith("tar.gz"))
-        or (tar_file_str.endswith("tgz"))
-        or (tar_file_str.endswith("tar"))
-    )
+    return tarfile.is_tarfile(tar_file)
 
 
 def extract_tar_file(tar_file: Path, dest_dir: Path = None) -> None:
