### 2023-01-26 7388df9
Update CONTRIBUTING.md
Co-authored-by: Alex Batisse <alexandre.batisse@hey.com>
