### 2023-01-26 e5d166f
Update CONTRIBUTING.md
Co-authored-by: Alex Batisse <alexandre.batisse@hey.com>
