### 2023-01-26 29dd698
update pynteny-dev.yml to include mkdocs
