### 2023-01-27 8a8d525
change default config / data dir
