+++ docs/references.md
@@ -1,1 +1,0 @@
-::: pynteny.filter
