### 2023-01-31 #67 @Robaina
Hi @Batalex,

thanks again for your review. I noticed that some of your comments were not addressed in the previous PRs.  I'll clarify some of them here (preparing a PR for the remaining):

1. `There are a few discrepancies in types, some of them could be fixed by converting argparse.ArgumentParser instances to CommandArgs before using them`

I understand what you mean here: Subcommand functions in `pynteny.subcommands` currently accept both an `argparse.ArgumentParser` object or a `pynteny.utils.CommandArgs` object. This flexible type in the argument was made on purpose so subcommand could be called both from the CLI and from within Python. I don't understand why this is undesirable. I could, like you pointed out transform argparse objects to CommandArgs, but find it unnecessary (unless there is a good reason behind that I'm missing).

Additionally, type hints in  `pynteny.subcommands`  could be fixed using  [typing's union](https://docs.python.org/3/library/typing.html#typing.Union)

What do you think?

2. `The defaults paths are quite inconvenient if I run the cli from the installed package rather than the source directory, the config file as well as the downloaded database ends up in /home//miniconda3/envs/Pynteny/lib/python3.10/site-packages/`

This has been changed already in this [PR](https://github.com/Robaina/Pynteny/pull/51).

3. `Since the author is using pathlib, they might use / or Path.joinpath instead of os.path.join`

This was changed in this [PR](https://github.com/Robaina/Pynteny/pull/43/files#diff-81e85ee580919075116db020decd73af951e180aef609fb320565e1030df7f63)

4. `I have not profiled the code to see if there is a performance bottleneck there, but I would use itertuples instead of iterrows in addHMMmetaInfoToHits. itertuples is much faster if you need to iterate over data frames one row at a time`

This was changed in this [PR](https://github.com/Robaina/Pynteny/pull/43/files#diff-0c49a2cbed47b6980cf9a8db2f1f5dfff83f7dbdd7c5cfff6b4d6eef24b09273)

(Will get rid of `getattr` in the next PR)

4. `My personal recommendation is to define the text encoding whenever a text file descriptor is opened. Even if this is not feasible at the moment, this is a tiny step toward Windows compatibility, and it removes a little bit of uncertainty.`

Alright, will add that in the new PR.

6. `The MG1655.gb file is stored in the tests folder, which is not included in the package. Therefore, when the package is installed using conda, it is not downloaded.`

Right, will modify this in the new PR.
