### 2023-01-31 3a3f17d
add utf encoding when openng files
