### 2023-01-31 ba5195b
using name attr in itertuples
