+++ src/pynteny/filter.py
@@ -398,17 +398,19 @@ class SyntenyHits:
         self._synteny_hits[fields] = ""
         # for i, row in self._synteny_hits.iterrows():
         for row in self._synteny_hits.itertuples():
-            i = getattr(row, "Index")
-            hmm_group = getattr(row, "hmm")
+            i = row.Index  # getattr(row, "Index")
+            # hmm_group = row.hmm  # getattr(row, "hmm")
             meta_values = [
                 [
                     str(v).replace("nan", "")
                     for k, v in pgap.get_meta_info_for_HMM(hmm).items()
                     if k != "#ncbi_accession"
                 ]
-                for hmm in hmm_group.split("|")  # row.hmm.split("|")
+                for hmm in row.hmm.split("|")  # hmm_group.split("|")
             ]
-            self._synteny_hits.loc[i, fields] = ["|".join(v) for v in zip(*meta_values)]
+            self._synteny_hits.loc[row.Index, fields] = [
+                "|".join(v) for v in zip(*meta_values)
+            ]
         return SyntenyHits(self._synteny_hits)
 
     def write_to_TSV(self, output_tsv: Path) -> None:
