### 2023-01-31 5639da2
itertuples
