+++ src/pynteny/filter.py
@@ -396,17 +396,14 @@ class SyntenyHits:
             return self._synteny_hits
         pgap = PGAP(hmm_meta)
         self._synteny_hits[fields] = ""
-        # for i, row in self._synteny_hits.iterrows():
         for row in self._synteny_hits.itertuples():
-            i = row.Index  # getattr(row, "Index")
-            # hmm_group = row.hmm  # getattr(row, "hmm")
             meta_values = [
                 [
                     str(v).replace("nan", "")
                     for k, v in pgap.get_meta_info_for_HMM(hmm).items()
                     if k != "#ncbi_accession"
                 ]
-                for hmm in row.hmm.split("|")  # hmm_group.split("|")
+                for hmm in row.hmm.split("|")
             ]
             self._synteny_hits.loc[row.Index, fields] = [
                 "|".join(v) for v in zip(*meta_values)
