### 2023-01-31 #67 @Batalex
Thanks, @Robaina; I needed to make a second pass at my comment, but you have already taken care of it!

Aside from the remaining `getattr` to take care of, I am pleased to see that you found interest in my comments.
While we wait for the second review, can I make some PR to add a few quality tools to ease the maintenance in the future?

I am thinking of:

- adding [isort](https://pycqa.github.io/isort/) to sort imports consistently
- adding [flake8](https://flake8.pycqa.org/en/latest/), a very common linter in the python ecosystem
- if you are feeling fancy, using [ruff ](https://github.com/charliermarsh/ruff) which can replace the two above
- adding [pre-commit](https://pre-commit.com/) hooks for black, isort, flake8 / ruff
- adding CI jobs to check the tools above before merging a PR
