### 2023-01-31 e692fcd
update docstrings
