### 2023-01-31 #67 @Robaina
Sure @Batalex, go ahead with this(these) PR(s). I'll be glad to help. These changes should make the codebase very robust and pretty.
