+++ src/pynteny/app/components.py
@@ -14,7 +14,6 @@ __author__ = meta["Author"]
 
 
 def show_sidebar():
-
     st.sidebar.image(
         st.session_state.sidebar_icon,
         use_column_width=True,
@@ -53,7 +52,6 @@ def show_sidebar():
             )
 
     with st.sidebar.expander("Advanced parameters:", expanded=False):
-
         st.markdown("Select custom HMM database")
         col1, col2 = st.columns([1, 1])
         with col1:
@@ -91,7 +89,6 @@ def show_sidebar():
 
 
 def show_mainpage():
-
     st.title("Pynteny — Synteny-aware HMM searches made easy")
 
     with st.expander("Select sequence data", expanded=False):
