### 2023-02-02 c806c27
refactor: Fix linter issues
