+++ src/pynteny/__init__.py
@@ -1,2 +1,4 @@
-from pynteny.api import Command, Search, Build, Download
+from pynteny.api import Build, Command, Download, Search
 from pynteny.cli import main
+
+__all__ = ("main", "Command", "Search", "Build", "Download")
