### 2023-02-02 aa695d9
chore: Add pre-commit config
