+++ .pre-commit-config.yaml
@@ -1,0 +1,12 @@
+default_stages: [ commit, push ]
+fail_fast: false
+repos:
+  - repo: https://github.com/psf/black
+    rev: 23.1.0
+    hooks:
+      - id: black
+        language_version: python3
+  - repo: https://github.com/charliermarsh/ruff-pre-commit
+    rev: 'v0.0.240'
+    hooks:
+      - id: ruff
