### 2023-02-03 0ebd0b9
refactor: Remove leftover Path creation
