### 2023-02-03 b3a5f9e
ci: Add linter + black checks to CI
