### 2023-02-17 bfbcd13
add file name to files in fasta dir
