+++ src/pynteny/preprocessing.py
@@ -89,6 +89,52 @@ def is_legit_DNA_sequence(record_seq: str) -> bool:
     return seq_symbols.issubset(nts)
 
 
+class FASTAmerger:
+    def __init__(self, input_dir: Path):
+        self.input_dir = Path(input_dir)
+
+    def prepend_filename_to_record_names(self, output_dir: Path) -> None:
+        """Prepend file name to each record label in fasta file
+
+        Args:
+            output_dir (Path): _description_
+        """
+        output_dir = Path(output_dir)
+        output_dir.mkdir(parents=True, exist_ok=True)
+        for file in self.input_dir.iterdir():
+            fasta = pyfastx.Fasta(file.as_posix(), build_index=False, full_name=True)
+            new_file_content = [
+                f">{file.stem}_{record_name}\n{record_seq}\n"
+                for record_name, record_seq in fasta
+            ]
+            with open(output_dir / file.name, "w+", encoding="UTF-8") as outfile:
+                outfile.writelines(new_file_content)
+
+    def merge(self, output_file: Path = None, prepend_file_name: bool = False) -> None:
+        """Merge input fasta files into a one (multi)fasta file.
+
+        Args:
+            output_file (Path, optional): path to ouput merged fasta file. Defaults to None.
+            prepend_file_name (bool, optional): whether to add file name as genome ID to
+            each record in the result merged fasta file.
+
+        """
+        if output_file is None:
+            output_file = self.input_dir / "merged.fasta"
+        else:
+            output_file = Path(output_file)
+        logger.info("Merging FASTA files in input directory")
+        # cmd_str = f'awk 1 * > {output_file}'
+        # # cmd_str = "find . -maxdepth 1 -type f --exec cat {} + > " + f"{output_file}"
+        cmd_str = f"printf '%s\\0' * | xargs -0 cat > {output_file}"
+        if prepend_file_name:
+            with tempfile.TemporaryDirectory() as tempdir:
+                self.prepend_filename_to_record_names(output_dir=tempdir)
+                utils.terminal_execute(cmd_str, work_dir=tempdir)
+        else:
+            utils.terminal_execute(cmd_str, work_dir=self.input_dir)
+
+
 class FASTA:
     """Handle and process fasta files."""
 
@@ -117,12 +163,16 @@ class FASTA:
         self._input_file = Path(new_path)
 
     @classmethod
-    def from_FASTA_directory(cls, input_dir: Path, merged_fasta: Path = None) -> FASTA:
+    def from_FASTA_directory(
+        cls, input_dir: Path, merged_fasta: Path = None, prepend_file_name: bool = False
+    ) -> FASTA:
         """Initialize FASTA class from directory of fasta files.
 
         Args:
             input_dir (Path): path to input directory.
             merged_fasta (Path, optional): path to output merged fasta. Defaults to None.
+            prepend_file_name (bool, optional): whether to add file name as genome ID to
+            each record in the result merged fasta file.
 
         Returns:
             FASTA: an initialized instance of class FASTA.
@@ -132,27 +182,8 @@ class FASTA:
             merged_fasta = input_dir / "merged_database.fasta"
         else:
             merged_fasta = Path(merged_fasta)
-        FASTA.merge(input_dir, merged_fasta)
+        FASTAmerger(input_dir).merge(merged_fasta, prepend_file_name)
         return cls(merged_fasta)
-
-    @staticmethod
-    def merge(input_dir: Path, output_file: Path = None) -> None:
-        """Merge input fasta files into a one (multi)fasta file.
-
-        Args:
-            input_dir (Path): path to input directory
-            output_file (Path, optional): path to ouput merged fasta file. Defaults to None.
-        """
-        input_dir = Path(input_dir)
[…]
