+++ docs/subcommands/build.md
@@ -11,9 +11,10 @@ usage: pynteny build [-h] [args]
 |short|long|default|help|
 | :--- | :--- | :--- | :--- |
 |`-h`|`--help`||show this help message and exit|
-|`-i`|`--data`|`None`|path to assembly input nucleotide data or annotated GenBank file.  It can be a single file or a directory of files (either of FASTA or GeneBank format).|
-|`-o`|`--outfile`|`None`|path to output (labeled peptide database) file. Defaults to file in the directory of input data.|
-|`-n`|`--processes`|`None`|set the maximum number of processes. Defaults to all but one.|
+|`-i`|`--data`|`None`|path to assembly input nucleotide data or annotated GenBank file.  It can be a single file or a directory of files (either of FASTA or GeneBank format).  If a directory, file name is prepended to the label of each translated peptide  originally coming from that file (i.e., to track the genome of origin)|
+|`-p`|`--prefix`|`None`|prefix to be added to each sequence obtained from original  input assembly or GeneBank file: e.g. genome accession.|
+|`-o`|`--outfile`|`None`|path to output (labelled peptide database) file. Defaults to  file in directory of input data.|
+|`-n`|`--processes`|`None`|set maximum number of processes. Defaults to all but one.|
 |`-l`|`--log`|`None`|path to log file. Log not written by default.|
 
 
