+++ envs/pynteny-dev.yml
@@ -27,3 +27,4 @@ dependencies:
     - mkdocs-jupyter
     - mkdocstrings[python]
     - ruff
+    - argmark
