### 2023-02-22 #67 @c-thoben
Here's my review! @Robaina @arianesasso

Thanks for onboarding me :) I am not that experienced, but I enjoyed looking into the review process.
You did some really nice work @Robaina and I hope my comments are helpful.

Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide
- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor before starting your review).

- [x] *A statement of need* clearly stating problems the software is designed to solve and its target audience in README.
- [x] *Installation instructions:* for the development version of the package and any non-standard dependencies in README.
- [x] *Vignette(s)* demonstrating major functionality that runs successfully locally.
- [x] *Function Documentation:* for all user-facing functions.
- [x] *Examples* for all user-facing functions.
- [x] *Community guidelines* including contribution guidelines in the README or CONTRIBUTING.
- [x] *Metadata* including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a pyproject.toml file or elsewhere.

- [x] Package has a README.md file in the root directory.The README should include, from top to bottom:

	- [x] A repostatus.org badge,

NOTE: If the README has many more badges, you might want to consider using a table for badges: see this example. Such a table should be more wide than high. (Note that the a badge for pyOpenSci peer-review will be provided upon acceptance.)

Reviewers are encouraged to submit suggestions (or pull requests) that will improve the usability of the package as a whole.Package structure should follow general community best-practices. In general please consider whether:

- [ ] *Installation:* Installation succeeds as documented. -> I had some trouble installing the software via conda. I have described them below.
- [x] *Functionality:* Any functional claims of the software been confirmed.
- [x] *Performance:* Any performance claims of the software been confirmed.
- [x] *Automated tests:* Tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
- [x] *Continuous Integration:* Has continuous integration setup (We suggest using Github actions but any CI platform is acceptable for review)
- [x] *Packaging guidelines*: The package conforms to the pyOpenSci packaging guidelines.
	- [x] Package supports modern versions of Python and not End of life versions.

- [x] The package has an *obvious research application* according to JOSS's definition in their submission requirements.
Note: Be sure to check this carefully, as JOSS's submission requirements and scope differ from pyOpenSci's in terms of what types of packages are accepted.

The package contains a paper.md matching JOSS's requirements with:
- [x] *A short summary* describing the high-level functionality of the software
- [x] *Authors:* A list of authors with their affiliations
- [x] *A statement of need* clearly stating problems the software is designed to solve and its target audience.
- [x] *References:* With DOIs for all those that have one (e.g. papers, datasets, software).

- [ ] *The author has responded to my review and made changes to my satisfaction. I recommend approving this package.

*Estimated hours spent reviewing:
10 h

---
- When trying to install the software, I had some issues with anaconda resolving the package dependencies if i used the recommended command "conda install -c pynteny". I got some help from the community and with "conda create -n pynteny -c conda-forge -c bioconda python pynteny" the installation worked fine. To test Pynteny, I used a VM with Ubuntu 22.04.1 LTS. The anaconda version was 23.1.10 and python3 inside the pynteny environment 3.10.6. If you need furhter information I am happy to provide them.
-  I have seen that if the PGAP database is downloaded, it is marked in the configuration file and the download is skipped at future exections. This can lead to problems if e.g. the user has deleted the database locally and wants to redownload it. Maybe a force option can be helpful.
-   All tests passed for me. However, I have seen that in the scripts "test_database_build.py" and "test_integration_search.py", the unittest main is not called so the tests are not executed if the module is run locally. Is this right?
- I enjoyed the quotes in the CLI :)
