### 2023-02-23 #67 @Robaina
Thank you so much for your review @c-thoben! I really appreciate the time you have taken to do this review.

Find below my answers to your comments:

> * When trying to install the software, I had some issues with anaconda resolving the package dependencies if i used the recommended command "conda install -c pynteny". I got some help from the community and with "conda create -n pynteny -c conda-forge -c bioconda python pynteny" the installation worked fine. To test Pynteny, I used a VM with Ubuntu 22.04.1 LTS. The anaconda version was 23.1.10 and python3 inside the pynteny environment 3.10.6. If you need furhter information I am happy to provide them.

The recommended way to install pynteny is via conda, in its own environment:

```bash
conda create -n pynteny
conda activate pynteny
conda install -c bioconda pynteny
```

as indicated in the [README](https://github.com/Robaina/Pynteny#readme). It looks like this is what you did at the end. Python should be included in the environment by default.

I have to update the package in bioconda. There have been a lot of changes in pynteny's codebase since then, but did not want to bother the bioconda community with further updates until this review was over.

The new features can be tested installing the developers version, following instructions in the [contributing](https://github.com/Robaina/Pynteny/blob/main/CONTRIBUTING.md#setting-up-a-local-developer-environment) file.

> * I have seen that if the PGAP database is downloaded, it is marked in the configuration file and the download is skipped at future exections. This can lead to problems if e.g. the user has deleted the database locally and wants to redownload it. Maybe a force option can be helpful.

I had the exact same thought. This [feature is already included](https://robaina.github.io/Pynteny/subcommands/download/) in the newer version of Pynteny (not yet in bioconda).

> * All tests passed for me. However, I have seen that in the scripts "test_database_build.py" and "test_integration_search.py", the unittest main is not called so the tests are not executed if the module is run locally. Is this right?

Right, I missed that. Thanks! It's been changed.

> * I enjoyed the quotes in the CLI :)

I'm glad you enjoyed the quotes :D
