### 2023-02-23 eaf3154
fixed issue with repeated hmms in structure
