+++ src/pynteny/api.py
@@ -173,7 +173,7 @@ class Build(Command):
     def __init__(
         self,
         data: Path,
-        prefix: str = None,
+        prepend: bool = False,
         outfile: Path = None,
         logfile: Path = None,
         processes: int = None,
@@ -190,7 +190,7 @@ class Build(Command):
 
         self._args = CommandArgs(
             data=Path(data),
-            prefix=prefix if prefix is not None else prefix,
+            prepend=prepend,
             outfile=Path(outfile) if outfile is not None else outfile,
             logfile=Path(logfile) if logfile is not None else logfile,
             processes=processes,
