### 2023-02-24 5a8b95d
added seqkit to env and readme
