### 2023-02-24 #67 @c-thoben
Hey @Robaina, thanks for your answer! It seems that most of my comments are already resolved. Onto the installation issue. I had the installation problems when I followed the recommended way in the README:
```
conda create -n pynteny
conda activate pynteny
conda install -c bioconda pynteny
```
The error I got is due to conda not being able to resolve the dependencies. The user who helped me had the same incompatibility on fedora37 shipping python3.11 with conda environments with python 3.10, 3.9 and 3.8. She found out that the installation works if the dependencies are resolved by conda at the same time:
 ```
conda create -n pynteny
conda activate pynteny
conda install -c conda-forge -c bioconda python pynteny
```
If you need more informations I am happy to provide them!
