### 2023-02-26 #67 @Robaina
Hi @c-thoben,

thank you so much for spotting and clarifying this issue. I have added some lines to README to point out this issue: try your solution and also install with mamba instead of conda (which is usually much faster to resolve dependencies). This should be enough for now. However, I'm thinking of further restricting dependency versions in the next update on bioconda, this should also help reduce installation time / solving dependencies.
