### 2023-02-27 #67 @lwasser
@Robaina i'm using a pure mamba forge installation and this install is not working for me on a MAC in a clean environment. something is up with the dependencies that i think needs to be resolved in a way that users can understand.

@c-thoben had a difficult time installing on linux as well and her points need to be addressed!

can you explain a few things to us?

1. conda install -c conda-forge -c bioconda python pynteny

why do you need to specify both conda-forge and bioconda channels when installing? are there some specific dependencies that are getting in the way?

2. the instructions in the readme are confusing because there are now 3 installation options as follows: there should be a single way to install a package that works.

```
Pynteny can be install with the following command:

conda install -c bioconda pynteny
If conda takes a long time to solve the dependencies (e.g., more than a few minutes) or fails to solve it, you may try:

conda install -c conda-forge -c bioconda python pynteny
Or install it with [mamba](https://github.com/mamba-org/mamba) instead of conda:

mamba install -c bioconda pynteny
```

finally there is a [personal channel in your environment here. ](https://github.com/Robaina/Pynteny/blob/main/envs/pynteny-dev.yml#L5). we want to ensure that authors are in fact using public channels. can please explain why a personal channel is listed in the environment?

Many thanks. It's critical that you address this installation issue as so many of us are running into confusion installing this package. While i understand the complexity of dealing with specific channels and complex installs, our goal here it so support usability in all of our vetted packages.
