### 2023-02-27 #67 @Robaina
Hi @lwasser,

thanks for your comments.

> @Robaina i'm using a pure mamba forge installation and this install is not working for me on a MAC in a clean environment. something is up with the dependencies that i think needs to be resolved in a way that users can understand.

I agree. For what @c-thoben  and you tell me, it seems like something is wrong with the package (now outdated) version in bioconda. Pynteny is built and installed without problems in the [tests](https://github.com/Robaina/Pynteny/blob/main/.github/workflows/tests.yml) GitHub Action, both in Ubuntu and Mac.

> can you explain a few things to us?
>
> 1. conda install -c conda-forge -c bioconda python pynteny
>
> why do you need to specify both conda-forge and bioconda channels when installing? are there some specific dependencies that are getting in the way?

bioconda channel needed for pynteny, conda-forge channel needed for python (solution suggested by @c-thoben). This was meant as a provisional fix until the new version was ready in bioconda. I have no trouble installing pynteny from bioconda in my Ubuntu 20.04 machine, so really could not know that this fix wouldn't work in other settings.

> 2. the instructions in the readme are confusing because there are now 3 installation options as follows: there should be a single way to install a package that works.
>
> ```
> Pynteny can be install with the following command:
>
> conda install -c bioconda pynteny
> If conda takes a long time to solve the dependencies (e.g., more than a few minutes) or fails to solve it, you may try:
>
> conda install -c conda-forge -c bioconda python pynteny
> Or install it with [mamba](https://github.com/mamba-org/mamba) instead of conda:
>
> mamba install -c bioconda pynteny
> ```

I agree. Ideally only ```conda install -c bioconda pynteny```, perhaps suggesting usage of mamba only to accelerate installation. Will update pynteny on bioconda and try installing it in the GitHub Action (Ubuntu and Mac) directly from bioconda once updated as a way to check installation on MacOS (I don't have access to a Mac machine)

>
> finally there is a [personal channel in your environment here. ](https://github.com/Robaina/Pynteny/blob/main/envs/pynteny-dev.yml#L5). we want to ensure that authors are in fact using public channels. can please explain why a personal channel is listed in the environment?
>

Thanks for spotting this. This channel was a remanent of a previous version in which pynteny was hosted in my personal conda channel. It was [requested](https://github.com/pyOpenSci/software-submission/issues/65#issuecomment-1352803647) for me to upload pynteny to bioconda prior to review. Already removed.

> Many thanks. It's critical that you address this installation issue as so many of us are running into confusion installing this package. While i understand the complexity of dealing with specific channels and complex installs, our goal here it so support usability in all of our vetted packages.

Of course. I totally agree. Thanks for raising these issues. I will update pynteny in bioconda, test installation again, and come back to you all.
