### 2023-02-27 bd6a4da
remove my channel from env
