+++ envs/pynteny-dev.yml
@@ -2,7 +2,6 @@ name: pynteny-dev
 channels:
   - defaults
   - bioconda
-  - semidan.robaina
   - conda-forge
 dependencies:
   - python >= 3.8
