### 2023-02-27 Robaina/Pynteny#76 @Robaina
Update pynteny in bioconda and check installation
There have been [reports](https://github.com/pyOpenSci/software-submission/issues/67#issuecomment-1444627694) of conda failing to solve dependencies when installing from bioconda.  Conda also seems to fail at solving dependencies in [MacOS](https://github.com/pyOpenSci/software-submission/issues/67#issuecomment-1446572212).

Check if updating pynteny in bioconda solves the issue with the dependencies. Further restricting conda package versions in the environment.yml may also help resolve dependencies (perhaps using conda lock)
