+++ src/pynteny/app/components.py
@@ -115,17 +115,17 @@ def show_mainpage():
     with st.expander("Enter synteny structure:", expanded=False):
         st.info(
             """
-            Synteny blocks are specified by strings of ordered HMM names or gene IDs with the following format: 
+            Synteny blocks are specified by strings of ordered HMM names or gene IDs with the following format:
 
-            $$\lt HMM_a \space n_{ab} \space \lt HMM_b \space n_{bc} \space \lt(HMM_{c1}|HMM_{c2}|HMM_{c3}),$$ 
+             $$\lt HMM_a \space n_{ab} \space \lt HMM_b \space n_{bc} \space \lt(HMM_{c1}|HMM_{c2}|HMM_{c3}),$$
 
-            where $n_{ab}$ corresponds to the maximum number of genes between $HMM_a$ and $HMM_b$. Additionally:
+             where $n_{ab}$ corresponds to the maximum number of genes between $HMM_a$ and $HMM_b$. Additionally:
 
-            - Results can be strand-specific, in that case $>$ preceding a HMM name indicates that the corresponding ORF must be located in the positive (or sense) strand. Likewise, a $<$ symbol indicates that the ORF must be located in the negative (antisense) strand. 
+             - Results can be strand-specific, in that case $>$ preceding a HMM name indicates that the corresponding ORF must be located in the positive (or sense) strand. Likewise, a $<$ symbol indicates that the ORF must be located in the negative (antisense) strand.
 
-            - Searches can be made strand-insensitive by omitting the $>$ or $<$ symbol. 
+             - Searches can be made strand-insensitive by omitting the $>$ or $<$ symbol.
 
-            - Several HMMs can be assigned to the same ORF, in which case the search is performed for all of them. In this case, HMM names must be separated by "|" and grouped within parentheses, as shown above.
+             - Several HMMs can be assigned to the same ORF, in which case the search is performed for all of them. In this case, HMM names must be separated by "|" and grouped within parentheses, as shown above.
             """
         )
 
