### 2023-02-27 #67 @Robaina
@lwasser

Sorry, I just realized. There is still one dependency, Streamlist-Aggrid, that is hosted in my [conda channel](https://anaconda.org/semidan.robaina/streamlit-aggrid). That's why my channel was still listed in the environment.

This is a non-critical dependency, it is employed to make an interactive results table in the web application. While installable [from pip](https://pypi.org/project/streamlit-aggrid/), I had to upload it to conda to be able to list it as a dependency in bioconda (pip dependencies not allowed).

Since this dependency is not hosted in a public channel, I will first try to contact Streamlit-Aggrid developer to suggest uploading it to bioconda, if unsuccessful, I will remove this dependency altogether and make a static results table in the web application.
