### 2023-02-27 #67 @lwasser
no worries @Robaina -Many thanks for your prompt reply.  i know that this review has been a LOT of work and you've been incredibly committed throughout the process! I just want us to figure out the install items and we can do that together if need be! just to be sure i've pinged a colleague who works on conda-forge to help us sort this out as well. but it ALSO sounds like part of this issue are things you've fixed that haven't been released yet which is very understandable.

i have seen packages (such as geopandas) that did have a dependency that you had to install manually. but i think it's not ideal for users. perhaps there is some way for you to provide instructions for using the interactive approach (if you wish) if you can't find a way to have this on conda.

however, If the package is on pypi already, it's actually quite simple to move it to [conda forge using greyskull](https://conda-forge.org/blog/posts/2020-03-05-grayskull/). please let us know if you need resources to share with the authors. i think this is quite doable and might then be less work for you assuming the authors are ok with it. then every time they deploy to pypi they also merge a pr on conda forge a few hours later. Then it's in both places. i've done this with several packages and it's easy to maintain.
