### 2023-02-27 #67 @Robaina
Thank you for the support, @lwasser. I'm sure we will figure this out.

Alright, so I went to Streamlit-Aggrid's repo and found out that I had already contacted the authors in October 2022 (so many things going on, I totally forgot).  Would be nice if we could set the conda-forge pipeline if the author agrees (contacted him again). I let you all know.
