### 2023-02-28 #67 @Robaina
Hi @lwasser,

I've been thinking about removing the Streamlit web application from pynteny and making it a separate pip-installable package `pynteny-app`  which would contain this functionality. This package would work as an optional add-on to pynteny (which would not be submitted to PyOpenSci)

This change would solve the issue with Streamlit-Aggrid, and probably also help solve dependencies by conda. All the functionality would still be accessible via CLI and within Python as a module (which are the main ways to access pynteny, anyway)

Would this change be acceptable with respect to the status of the current submission?

Thanks!
