### 2023-02-28 #67 @lwasser
@Robaina i think so. but let's check in with @arianesasso and our reviewers  @Batalex @c-thoben so they are aware of the update. i see that this dependency is really not core to your functionality so it could be a nice add on as you mention and would streamline your installation significantly! Ari, Alex and Corinna - does this seem reasonable? i don't want to take over this review i just want to help ensure things are installable. many thanks. i can stop complicatin gthe review once we have decided :)
