### 2023-03-01 ae72fdc
fixed hmm patter issue Check HMM name retrieval from synteny structura #78
