### 2023-03-01 #67 @lwasser
it actually was already there - https://anaconda.org/conda-forge/streamlit-aggrid @Robaina can you confirm that this is what you need? if so it should just be a simple environment update to make everything work.
