+++ src/pynteny/app/components.py
@@ -1,155 +1,0 @@
-import os
-from importlib import metadata
-
-import streamlit as st
-
-import pynteny.app.callbacks as callbacks
-import pynteny.app.filemanager as filemanager
-from pynteny.app.helpers import plot_dataframe
-
-meta = metadata.metadata("pynteny")
-__version__ = meta["Version"]
-__author__ = meta["Author"]
-
-
-def show_sidebar():
-    st.sidebar.image(
-        st.session_state.sidebar_icon,
-        use_column_width=True,
-        caption=f"Synteny-aware HMM searches made easy (v{__version__})",
-    )
-
-    st.sidebar.text(" ")
-    st.sidebar.text(" ")
-
-    with st.sidebar.expander("Current output directory:", expanded=True):
-        if (st.session_state.outdir is not None) and (st.session_state.outdir.exists()):
-            files_div = st.empty()
-            with files_div.container():
-                filemanager.show_files_in_dir(st.session_state.outdir, sidebar=False)
-
-        st.text(" ")
-        st.button("Select directory", on_click=callbacks.update_output_dir)
-        col1, col2 = st.columns([0.5, 0.5])
-        with col1:
-            st.text_input(
-                "",
-                value="",
-                max_chars=None,
-                key="subdirectory",
-                on_change=callbacks.update_output_subdirectory,
-                placeholder="Create subdirectory",
-            )
-        with col2:
-            st.text_input(
-                "",
-                value="",
-                max_chars=None,
-                key="prefix",
-                on_change=callbacks.update_output_prefix,
-                placeholder="Enter output prefix",
-            )
-
-    with st.sidebar.expander("Advanced parameters:", expanded=False):
-        st.markdown("Select custom HMM database")
-        col1, col2 = st.columns([1, 1])
-        with col1:
-            st.button("HMM directory", on_click=callbacks.select_HMM_dir)
-        with col2:
-            st.button("HMM metadata", on_click=callbacks.select_HMM_meta)
-
-        col1, col2, col3 = st.columns([0.4, 0.3, 0.3])
-        with col1:
-            st.markdown(("Processes:"))
-            st.slider(
-                "Processes",
-                min_value=1,
-                max_value=os.cpu_count(),
-                value=os.cpu_count() - 1,
-                step=1,
-                on_change=callbacks.set_number_of_processes,
-                key="processes",
-            )
-        with col2:
-            st.markdown(("Output log:"))
-            st.select_slider(
-                "",
-                options=["No", "Yes"],
-                value="Yes",
-                on_change=callbacks.select_log_path,
-                key="log",
-            )
-    st.sidebar.text(" ")
-    col1, col2 = st.sidebar.columns([1, 1])
-    with col1:
-        st.button("Restart session")
-    with col2:
-        st.button("Close session", key="close", on_click=callbacks.close_session)
-
-
-def show_mainpage():
-    st.title("Pynteny — Synteny-aware HMM searches made easy")
-
-    with st.expander("Select sequence data", expanded=False):
-        st.info(
-            """
-            Sequence data can be either:
-
-            - nucleotide assembly data in FASTA format or
-            - a GenBank file containing sequence annotations.
-
-            __Note__: This Pynteny instance is run locally, thus files are always kept in your machine.
-            """
-        )
-
-    with st.expander("", expanded=True):
-        col1, col2 = st.columns([1, 1])
-        with col1:
-            st.button("Upload file", on_click=callbacks.upload_data)
-        with col2:
-            st.button("Build database", on_click=callbacks.build)
-
-    if st.session_state.sequence_data_uploaded:
-        st.info(f"Uploaded file: {st.session_state.search_state.data.name}")
[…]
