### 2023-03-02 #67 @Batalex
On the one hand, I think that the streamlit app is orthogonal to the package itself. I mean that removing it does not change the package's "elevator pitch."
On the other hand, the web app is mentioned in the `paper.md` (https://github.com/Robaina/Pynteny/blob/main/ms/paper.md) submitted to JOSS.

I would be +1 on moving it to another package, if it means that a larger pool of users can use the core package, but I wonder if it could be an issue with JOSS.
