### 2023-03-03 #67 @Robaina
Great, I agree. I will edit the paper. I should be fine since it is not even submitted yet. I'll let you know once I test installation after these new changes.
