+++ src/pynteny/cli.py
@@ -67,12 +67,12 @@ class Pynteny:
         print(
             (
                 """
-    ____              __                  
+    ____              __
    / __ \__  ______  / /____  ____  __  __
   / /_/ / / / / __ \/ __/ _ \/ __ \/ / / /
- / ____/ /_/ / / / / /_/  __/ / / / /_/ / 
-/_/    \__, /_/ /_/\__/\___/_/ /_/\__, /  
-      /____/                     /____/   
+ / ____/ /_/ / / / / /_/  __/ / / / /_/ /
+/_/    \__, /_/ /_/\__/\___/_/ /_/\__, /
+      /____/                     /____/
 
 """
                 f"Synteny-based Hmmer searches made easy, v{__version__}\n"
