+++ src/conda/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "pynteny" %}
-{% set version = "0.0.6" %}
+{% set version = "1.0.0" %}
 
 package:
   name: "{{ name|lower }}"
