+++ .github/workflows/test_build.yml
@@ -34,7 +34,7 @@ jobs:
       - name: create environment
         uses: conda-incubator/setup-miniconda@v2
         with:
-          python-version: 3.8
+          python-version: 3.9
           # mamba-version: "*"
           # channels: conda-forge,bioconda,semidan.robaina,defaults
           # auto-activate-base: false
