+++ .github/workflows/test_build.yml
@@ -33,8 +33,8 @@ jobs:
 
       - name: create environment
         uses: conda-incubator/setup-miniconda@v2
-        with:
-          python-version: 3.9
+        # with:
+        #   python-version: 3.9
           # mamba-version: "*"
           # channels: conda-forge,bioconda,semidan.robaina,defaults
           # auto-activate-base: false
