+++ .github/workflows/test_build.yml
@@ -33,16 +33,16 @@ jobs:
 
       - name: create environment
         uses: conda-incubator/setup-miniconda@v2
-        # with:
-        #   python-version: 3.9
+        with:
+          python-version: 3.10
           # mamba-version: "*"
-          # channels: conda-forge,bioconda,semidan.robaina,defaults
+          channels: conda-forge,bioconda,defaults
           # auto-activate-base: false
           # activate-environment: tests_pynteny
           # environment-file: envs/pynteny-dev.yml
 
       - name: Build & Install Pynteny
-        run: conda install -c bioconda pynteny
+        run: conda install pynteny
 
       - name: Run tests and collect coverage
         run: pynteny --help && python -m unittest discover tests
