+++ .github/workflows/test_build.yml
@@ -34,7 +34,7 @@ jobs:
       - name: create environment
         uses: conda-incubator/setup-miniconda@v2
         with:
-          python-version: 3.10
+          python-version: "3.10"
           # mamba-version: "*"
           channels: conda-forge,bioconda,defaults
           # auto-activate-base: false
