+++ .github/workflows/conda_build.yml
@@ -3,6 +3,7 @@ name: publish_conda
 on:
   release:
     types: [published]
+  push:
     
 jobs:
   publish:
