+++ .github/workflows/conda_build.yml
@@ -11,7 +11,7 @@ jobs:
     steps:
     - uses: actions/checkout@v1
     - name: publish-to-conda
-      uses: MichaelsJP/conda-package-publish-action@v1.0.0
+      uses: MichaelsJP/conda-package-publish-action@v1.1.0
       with:
         subDir: '.conda'
         AnacondaToken: ${{ secrets.ANACONDA_TOKEN }}
