+++ .github/workflows/conda_build.yml
@@ -10,6 +10,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v1
+    - name: conda-channels
+      run: conda config --show channels
     - name: publish-to-conda
       uses: MichaelsJP/conda-package-publish-action@v1.1.0
       with:
