+++ .github/workflows/conda_build.yml
@@ -13,5 +13,5 @@ jobs:
     - name: publish-to-conda
       uses: maxibor/conda-package-publish-action@v1.1
       with:
-        subDir: 'conda'
+        subDir: '.conda'
         AnacondaToken: ${{ secrets.ANACONDA_TOKEN }}
