+++ .github/workflows/conda_build.yml
@@ -3,7 +3,6 @@ name: publish_conda
 on:
   release:
     types: [published]
-  push:
     
 jobs:
   publish:
