+++ .github/workflows/test_build.yml
@@ -34,7 +34,7 @@ jobs:
       - name: create environment
         uses: conda-incubator/setup-miniconda@v2
         with:
-          python-version: "3.10"
+          # python-version: "3.10"
           # mamba-version: "*"
           channels: conda-forge,bioconda,defaults
           # auto-activate-base: false
@@ -42,7 +42,7 @@ jobs:
           # environment-file: envs/pynteny-dev.yml
 
       - name: Build & Install Pynteny
-        run: conda install -c semidan.robaina pynteny
+        run: conda install -c conda-forge -c bioconda python=3.10 pynteny
 
       - name: Run tests and collect coverage
         run: pynteny --help && python -m unittest discover tests
