### 2023-03-06 #67 @Robaina
Hi there @arianesasso, @c-thoben @Batalex ,

I have updated pynteny on bioconda, and tested installation directly from bioconda within a [GitHub Action](https://github.com/Robaina/Pynteny/actions/runs/4342395513) (under ubuntu-lastest and macos-latest).

I also updated setup instructions in the README, and modified python version, since now pynteny depends on python 3.10.

Please, could you verify that you can install pynteny in your settings? (perhaps @lwasser could also try installing it in MacOS?)

Thanks!
