### 2023-03-06 #67 @lwasser
sure thing i'll try to install on macOS @c-thoben can you please try to install again too following the new readme instructions? @Robaina thank you for your dedication and promptness in getting all of these things sorted out and fixed! it's much appreciated.
