### 2023-03-06 #67 @lwasser
i can confirm the same issue and response on my mac!

whats odd is i think the error is it can't find pynteny on bioconda

````
PackagesNotFoundError: The following packages are not available from current channels:

  - pynteny
 ```

 @Robaina do you need help with this? i think we can get you some help. can i invite you to our slack? we can talk there? if so please email me at leah @ pyopensci.org and i'll send you an invite so we can try to figure this out together.
