### 2023-03-07 #67 @Robaina
Sure thing @lwasser, will send you an email for that.

I think the issue might be related to the "new" ARM64 processor architecture in MAC. The machine in which @Batalex tried to install pynteny seems to have one (and all the new apple machines starting November 2020, I think).

Several dependencies of pynteny don't provide support for ARM64, e.g. [hmmer](https://anaconda.org/bioconda/hmmer), [prodigal](https://anaconda.org/bioconda/prodigal), [pyfastx](https://anaconda.org/bioconda/pyfastx). Thus, pynteny itself doesn't have support for ARM64 architectures as of now. (To compare, here is another dependency [psutil](https://anaconda.org/conda-forge/psutil) which does have ARM64 support on bioconda).

If that's the case, there is a workaround, basically telling conda to construct an environment using x86 architecture on ARM64 machines to install dependencies not yet supporting ARM64. [Here](https://towardsdatascience.com/how-to-manage-conda-environments-on-an-apple-silicon-m1-mac-1e29cb3bad12) is a blog post with instructions, for what I read, something like:

```
CONDA_SUBDIR=osx-64 conda create -n pynteny_x86 python=3.10
conda activate pynteny_x86
conda config --env --set subdir osx-64
```

and then, I guess:

```
(pynteny_x86) conda install -c bioconda pynteny
```

That's something new I learned today... I have no way to test if this workaround will be sufficient, though. Will send you the email now, anyway.

Thanks!
