### 2023-03-07 #67 @Batalex
The joys of novelty 😬
