### 2023-03-07 #67 @c-thoben
@Robaina I am able to install Pynteny with the README instructions on my Linux VM and also on MacOS without problems. As I have a 2018 MacBook, this may support the ARM64 processor theory. Thanks for looking into this issue @Robaina and thanks for your support @lwasser!
