### 2023-03-07 #67 @lwasser
Ok i just installed in my M1 - the sequence below worked!
```
CONDA_SUBDIR=osx-64 conda create -n pynteny_x86 python=3.10
conda activate pynteny_x86
conda config --env --set subdir osx-64
conda install -c bioconda pynteny
```

it did take it a moment to import into python but it did import successfully
NOTE: i talked with a conda forge maintainer about this - IF those three packages on bioconda could be also added to conda-forge then they can actually be added to a [list where a bot tries ](https://github.com/conda-forge/conda-forge-pinning-feedstock/blob/main/recipe/migrations/osx_arm64.txt) to build  the appropriate arch wheels (arm64 for osx in this case). of course there might be some reason that they aren't already there so this could be a non trivial task. but i think for our purposes if the package installs on everyone's computer who is here in this review, i think we'd done due diligence. and all that is needed is a documentation update for how to install.
