### 2023-03-07 2046d32
update setup on README
