### 2023-03-07 #67 @Robaina
Excellent. Thanks for consulting this with conda-forge @lwasser . I have added these extra lines to README and also the documentation. This issue can always be updated in future releases.

Thanks so much everyone for the review!
