### 2023-03-07 #67 @lwasser
ok wonderful @Robaina @arianesasso will followup with next steps. so appreciative of all of the back and forth on this review. Essentially we just need to get an Ok from both reviewers that the changes to your package have satisfied their requests in the review... and then there are a few closing steps and we can send you on your way to JOSS! more soon...
