+++ envs/pynteny-dev.yml
@@ -10,7 +10,7 @@ dependencies:
   - hmmer >= 3.3
   - prodigal >= 2.6
   - pyfastx >=0.8
-  - biopython >= 1.8y
+  - biopython >= 1.8
   - numpy
   - pandas
   - seqkit
