### 2023-03-10 #67 @Robaina
Thank you @Batalex. Likewise, you far exceeded my expectations in your review. Incredibly supportive and collaborative. From now on using `ruff --fix` in all my projects 😆
