### 2023-03-10 #67 @Robaina
Thank you so much for your time and great review @c-thoben, and for helping me identify and solve Pynteny's installation problems!
