### 2023-03-10 #67 @lwasser
# Pynteny is now a part of the pyOpenSci Ecosystem!
----------------------------------------------
🎉 pynTeny has been approved by pyOpenSci! Thank you @Robaina  for submitting pyteny and many thanks to @Batalex @c-thoben for reviewing this package! 😸 NOTE:  I have to say this was an extraordinary review over all. @NickleDave you also really did a wonderful job in getting this review together early on. All aspects of this review went exactly as i would hope a pyOpenSci review would go!! We want our reviews to be a conversation that **supports** maintainers in improving their package! And i really think this was achieved here. So thank you ALL for being a tremendous team in this review.

There are a few things left to do to wrap up this submission after you've taken the post review survey!

## Author Tasks

@Robaina please do the following:
- [ ] Activate [Zenodo](https://zenodo.org/) watching the repo if you haven't already done so. (please ask if this if you have any questions about this!)
- [ ] Tag and create a release to create a Zenodo version and DOI. Please share the DOI here so either @arianesasso or myself can update the archive at the top of this issue!
- [ ] Add the badge for pyOpenSci peer-review to the README.md of pynteny. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.
- [ ] Add <package-name> to the pyOpenSci website. @Robaina , please open a pr to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml): to add your package and name to the list of contributors.

## Editor Tasks
- [ ] Please add everyone involved in this review to our contributors list using the all-contributors bot
_NOTE: unfortunately you will need to do this one contributor at a time and merge the pr that is opened each time._
Example comment to do this: `@all-contributors please add @robaina for code, review`

<IF JOSS SUBMISSION>

- [ ] Please add a link to the JOSS review issue once it is created here as well
- [ ] When the JOSS review is complete, add a comment to your review here in the pyOpenSci software-review repo that it has been approved by JOSS.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉
<IF JOSS SUBMISSION/>

All -- if you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and documentation in the [software-peer-review-guide](https://www.pyopensci.org/software-peer-review). We have also been updating our documentation to improve the process, so all feedback is appreciated!
