### 2023-03-10 #67 @NickleDave
>  @NickleDave you also really did a wonderful job in getting this review together early on.

It was really easy to work with @Robaina who has been really responsive to feedback and who has now evolved `pynteny` into a full-fledged package, happy to be a part of the review
