### 2023-03-11 #67 @Robaina
Thank you so much you all! I think I have completed all the required checks from my side. Please, review my [PR](https://github.com/pyOpenSci/pyopensci.github.io/pull/143) to include Pynteny in the database. Not sure I completed the fields correctly.

As a final note. I have to say that I have enjoyed this review process. You have been tremendously supportive. Feedback has been always constructive to better the package. I wished these kinds of open reviews were the norm in Science...
