### 2023-03-15 #67 @arianesasso
It is really great to know you enjoyed the process! I think we all did 😄 .
Now, for the last steps regarding JOSS:

[Please read their paper.md file requirements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), and make sure that the paper is in your repo.

Then, follow the instructions here: [submitting the package to JOSS](https://joss.readthedocs.io/en/latest/submitting.html).

Now, you don't need to go through another review 🙂 : You should just tag the JOSS issue with our issue.
