+++ src/pynteny/preprocessing.py
@@ -528,6 +528,7 @@ class GeneAnnotator:
         metagenome: bool = True,
         output_file: Path = None,
         prodigal_args: str = None,
+        tmpdir: Path = None,
     ) -> LabelledFASTA:
         """Run prodigal on assembly and export single fasta file with peptide ORFs predictions
 
@@ -536,6 +537,7 @@ class GeneAnnotator:
             metagenome (bool, optional): whether assembled sequences correspond to metagenomic data. Defaults to True.
             output_file (Path, optional): path to output fasta file. Defaults to None.
             prodigal_args (str, optional): additional arguments to be passed to prodigal CLI. Defaults to None.
+            tmpdir (Path, optional): path to temporary directory. Defaults to tempfile default.
 
         Returns:
             LabelledFASTA: object containing the labelled peptide database.
@@ -549,7 +551,17 @@ class GeneAnnotator:
             )
         else:
             output_file = Path(output_file)
-        with tempfile.TemporaryDirectory() as contigs_dir, tempfile.TemporaryDirectory() as prodigal_dir, tempfile.NamedTemporaryFile() as temp_fasta:
+        if tmpdir is None:
+            tmpdir = Path(tempfile.gettempdir())
+        else:
+            tmpdir = Path(tmpdir).resolve()
+
+        with tempfile.TemporaryDirectory(
+            dir=tmpdir / "contigs"
+        ) as contigs_dir, tempfile.TemporaryDirectory(
+            dir=tmpdir / "prodigal"
+        ) as prodigal_dir, tempfile.NamedTemporaryFile() as temp_fasta:
+
             contigs_dir = Path(contigs_dir)
             prodigal_dir = Path(prodigal_dir)
             logger.info("Running prodigal on assembly data")
@@ -635,6 +647,7 @@ class Database:
         prepend_file_name: bool = False,
         output_file: Path = None,
         processes: int = None,
+        tmpdir: Path = None,
     ) -> LabelledFASTA:
         """Build database from data files.
 
@@ -645,6 +658,7 @@ class Database:
                 each record in the result merged fasta file.
             output_file (Path, optional): path to output file. Defaults to None.
             processes (int, optional): maximum number of threads. Defaults to all minus one.
+            tmpdir (Path, optional): path to temporary directory. Defaults to tempfile default.
 
         Returns:
             LabelledFASTA: object containing the labelled peptide database.
@@ -667,6 +681,7 @@ class Database:
                 labelled_database = GeneAnnotator(assembly_fasta).annotate(
                     output_file=output_file,
                     processes=processes,
+                    tmpdir=tmpdir,
                 )
         elif self.is_gbk(self._data_files[0]):
             logger.info("Parsing GenBank data.")
