+++ src/pynteny/subcommands.py
@@ -182,6 +182,7 @@ def build_database(args: Union[CommandArgs, ArgumentParser]) -> None:
         output_file=args.outfile,
         prepend_file_name=prepend_file_name,
         processes=args.processes,
+        tmpdir=args.tmpdir,
     )
     logger.info("Database built successfully!")
     logging.shutdown()
