+++ src/pynteny/subcommands.py
@@ -182,7 +182,7 @@ def build_database(args: Union[CommandArgs, ArgumentParser]) -> None:
         output_file=args.outfile,
         prepend_file_name=prepend_file_name,
         processes=args.processes,
-        tmpdir=args.tmpdir,
+        tmpdir=args.tempdir,
     )
     logger.info("Database built successfully!")
     logging.shutdown()
