+++ docs/subcommands/build.md
@@ -12,9 +12,9 @@ usage: pynteny build [-h] [args]
 | :--- | :--- | :--- | :--- |
 |`-h`|`--help`||show this help message and exit|
 |`-i`|`--data`|`None`|path to assembly input nucleotide data or annotated GenBank file.  It can be a single file or a directory of files (either of FASTA or GeneBank format).  If a directory, file name is prepended to the label of each translated peptide  originally coming from that file (i.e., to track the genome of origin)|
-|`-p`|`--prepend-filename`||whether to prepend file name to peptide sequences when a directory with multiple fasta or genbank files is used as input.|
+|`-x`|`--prepend-filename`||whether to prepend file name to peptide sequences when a directory with multiple fasta or genbank files is used as input.|
 |`-o`|`--outfile`|`None`|path to output (labelled peptide database) file. Defaults to  file in directory of input data.|
-|`-n`|`--processes`|`None`|set maximum number of processes. Defaults to all but one.|
+|`-p`|`--processes`|`None`|set maximum number of processes. Defaults to all but one.|
 |`-t`|`--tempdir`|`None`|path to directory to store temporary files. Defaults to system temp directory.|
 |`-l`|`--log`|`None`|path to log file. Log not written by default.|
 
