+++ ms/paper.bib
@@ -192,3 +192,19 @@
 	keywords = {comparative genomics, cotton, grasses, orthology, sex chromosomes, synteny},
 	pages = {e78526},
 }
+
+@article{pgap,
+    author = {Li, Wenjun and O’Neill, Kathleen R and Haft, Daniel H and DiCuccio, Michael and Chetvernin, Vyacheslav and Badretdin, Azat and Coulouris, George and Chitsaz, Farideh and Derbyshire, Myra K and Durkin, A Scott and Gonzales, Noreen R and Gwadz, Marc and Lanczycki, Christopher J and Song, James S and Thanki, Narmada and Wang, Jiyao and Yamashita, Roxanne A and Yang, Mingzhang and Zheng, Chanjuan and Marchler-Bauer, Aron and Thibaud-Nissen, Françoise},
+    title = "{RefSeq: expanding the Prokaryotic Genome Annotation Pipeline reach with protein family model curation}",
+    journal = {Nucleic Acids Research},
+    volume = {49},
+    number = {D1},
+    pages = {D1020-D1028},
+    year = {2020},
+    month = {12},
+    abstract = "{The Reference Sequence (RefSeq) project at the National Center for Biotechnology Information (NCBI) contains nearly 200 000 bacterial and archaeal genomes and 150 million proteins with up-to-date annotation. Changes in the Prokaryotic Genome Annotation Pipeline (PGAP) since 2018 have resulted in a substantial reduction in spurious annotation. The hierarchical collection of protein family models (PFMs) used by PGAP as evidence for structural and functional annotation was expanded to over 35 000 protein profile hidden Markov models (HMMs), 12 300 BlastRules and 36 000 curated CDD architectures. As a result, \\&gt;122 million or 79\\% of RefSeq proteins are now named based on a match to a curated PFM. Gene symbols, Enzyme Commission numbers or supporting publication attributes are available on over 40\\% of the PFMs and are inherited by the proteins and features they name, facilitating multi-genome analyses and connections to the literature. In adherence with the principles of FAIR (findable, accessible, interoperable, reusable), the PFMs are available in the Protein Family Models Entrez database to any user. Finally, the reference and representative genome set, a taxonomically diverse subset of RefSeq prokaryotic genomes, is now recalculated regularly and available for download and homology searches with BLAST. RefSeq is found at https://www.ncbi.nlm.nih.gov/refseq/.}",
+    issn = {0305-1048},
+    doi = {10.1093/nar/gkaa1105},
+    url = {https://doi.org/10.1093/nar/gkaa1105},
+    eprint = {https://academic.oup.com/nar/article-pdf/49/D1/D1020/35364280/gkaa1105\_supplemental\_file.pdf},
+}
