+++ ms/paper.md
@@ -32,16 +32,31 @@ While function is generally conserved among sequence orthologs, i.e., homologs t
 
 # Statement of need
 
-Here we introduce `Pynteny`, a Python tool designed to conduct synteny-aware, profile HMM searches in prokaryotic sequence databases. `Pynteny` facilitates querying sequence databases with arrangements of profile HMMs that reflect a target syntenic block. To this end, it enables encoding positional information, such as gene order, maximum in-between gene distances, and strand specificity, into the search query. `Pynteny` is designed to work directly with assembled nucleotide sequence data, however, it also accepts annotated genomes in GenBank format as input data. It relies on Prodigal [@prodigal] to translate and add positional tags to individual genes, and on HMMER3, [@hmmer] to search sequence databases for homologs through profile HMMs.
+Here we introduce `Pynteny`, a Python tool designed to conduct synteny-aware, profile HMM searches in prokaryotic sequence databases. `Pynteny` facilitates querying sequence databases with arrangements of profile HMMs that reflect a target syntenic block. To this end, it enables encoding positional information, such as gene order, maximum in-between gene distances, and strand specificity, into the search query. 
 
-`Pynteny` was designed to be used by researchers working with large, unannotated sequence databases, such as those typically encountered in metagenomic analyses. It can be accessed through a command line interface or easily integrated into pipelines as a Python package.
+For instance, consider the following syntenic block:
+
+![synteny example.\label{fig:example}](synteny_example.png){ width=20% }
+
+The syntenic block shown above is composed of four genes, `A`, `B`, `C`, and `D`. Genes A-C locate consecutively in the positive strand, and are followed by three (untargeted) genes and by gene D, which is located in the negative strand. `Pynteny` allows searching for the syntenic block above with the following query string:
+
+$>HMM_A \:\: 0 \:\: >(HMM_{B1} | HMM_{B2}) \:\: 0 \:\: >HMM_C \:\: 3 \:\:< HMM_D,$
+
+where $HMM_a$ represents the name of the HMM modeling gene A, eacn integer represents the maximum number of genes between consecutive, and `<` and `>` indicate the strand in which to search for the HMM pattern, antisense and sense, respectively. Alternative HMMs can be used for a single gene, as shown in the HMM group $(HMM_{B1} | HMM_{B2})$ above, in which case the search will be performed with both HMMs. Additionally, gene symbols can be used directly in the query string when the [@pgap] HMM database is employed, and which is the default database used by `Pynteny`.
+
+`Pynteny` was designed to be used by researchers working with large, unannotated sequence databases, such as those typically encountered in metagenomic analyses. It can be accessed through a command line interface or easily integrated into pipelines as a Python package. It can directly handle assembled nucleotide sequence data, however, it also accepts annotated genomes in GenBank format as input data. In both cases, the package provides all the necessary functionality to preprocess the sequence database. 
+
+`Pynteny` relies on Prodigal [@prodigal] to translate and add positional tags to individual genes, and on HMMER3, [@hmmer] to search sequence databases for homologs through profile HMMs. Usage information as well as examples in the form of Jupyter Notebooks for both the command line interface and the Python package are available in the [documentation](https://robaina.github.io/Pynteny/).
 
 # State of the field
 
[…]
