+++ ms/paper.md
@@ -38,11 +38,11 @@ For instance, consider the following syntenic block:
 
 ![synteny example.\label{fig:example}](synteny_example.png){ width=20% }
 
-The syntenic block shown above is composed of four genes, `A`, `B`, `C`, and `D`. Genes A-C locate consecutively in the positive strand, and are followed by three (untargeted) genes and by gene D, which is located in the negative strand. `Pynteny` allows searching for the syntenic block above with the following query string:
+The syntenic block shown above is composed of four genes, `A`, `B`, `C`, and `D`. Genes A-C locate consecutively in the positive strand and are followed by three (untargeted) genes and by gene D, which is located in the negative strand. `Pynteny` allows searching for the syntenic block above with the following query string:
 
 $>HMM_A \:\: 0 \:\: >(HMM_{B1} | HMM_{B2}) \:\: 0 \:\: >HMM_C \:\: 3 \:\:< HMM_D,$
 
-where $HMM_a$ represents the name of the HMM modeling gene A, eacn integer represents the maximum number of genes between consecutive, and `<` and `>` indicate the strand in which to search for the HMM pattern, antisense and sense, respectively. Alternative HMMs can be used for a single gene, as shown in the HMM group $(HMM_{B1} | HMM_{B2})$ above, in which case the search will be performed with both HMMs. Additionally, gene symbols can be used directly in the query string when the [@pgap] HMM database is employed, and which is the default database used by `Pynteny`.
+where $HMM_a$ represents the name of the HMM modeling gene A, each integer represents the maximum number of genes between consecutive, and `<` and `>` indicate the strand in which to search for the HMM pattern, antisense and sense, respectively. Alternative HMMs can be used for a single gene, as shown in the HMM group $(HMM_{B1} | HMM_{B2})$ above, in which case the search will be performed with both HMMs. Additionally, gene symbols can be used directly in the query string when the [@pgap] HMM database is employed, which is the default database used by `Pynteny`.
 
 `Pynteny` was designed to be used by researchers working with large, unannotated sequence databases, such as those typically encountered in metagenomic analyses. It can be accessed through a command line interface or easily integrated into pipelines as a Python package. It can directly handle assembled nucleotide sequence data, however, it also accepts annotated genomes in GenBank format as input data. In both cases, the package provides all the necessary functionality to preprocess the sequence database. 
 
@@ -50,7 +50,7 @@ where $HMM_a$ represents the name of the HMM modeling gene A, eacn integer repre
 
 # State of the field
 
-Several existing tools are dedicated to the exploration, analysis, and visualization of synteny blocks among genomes. In these tools, users typically input a number of annotated genomes and obtain a collection of syntenic relations of shared gene sets among the genomes. Examples of these tools are MCScan [@mcscan] and MCScanX [@mcscanx], Clinker [@clinker], pyGenomeViz [@pyGenomeViz], genePlotR [@geneplotR], gggenomes [@gggenomes], GENESPACE [@genespace] and Mology [@mology]. These tools are excellent resources for the identification and analysis of syntenic relations among genomes, and they are functionally complementary to `Pynteny`. Specifically, rather than exploring syntenic blocks within annotated genomes, `Pynteny`'s objective is to search for specific syntenic structures within unannotated (assembled) sequence data, as well as to leverage syntenic information to reduce uncertainty due to paralogs during function annotation. Therefore, `Pynteny` requires a previous identification of conserved syntenic structures, which can be obtained from existing tools such as the ones previously indicated.
[…]
