+++ .github/workflows/conda_build.yml
@@ -1,16 +1,0 @@
-name: publish_conda
-
-on:
-  release:
-    types: [published]
-    
-jobs:
-  publish:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v1
-    - name: publish-to-conda
-      uses: maxibor/conda-package-publish-action@v1.1
-      with:
-        subDir: '.conda'
-        AnacondaToken: ${{ secrets.ANACONDA_TOKEN }}
