+++ ms/paper.md
@@ -42,7 +42,7 @@ The syntenic block shown above is composed of four genes, `A`, `B`, `C`, and `D`
 
 $>HMM_A \:\: 0 \:\: >(HMM_{B1} | HMM_{B2}) \:\: 0 \:\: >HMM_C \:\: 3 \:\:< HMM_D,$
 
-where $HMM_a$ represents the name of the HMM modeling gene A, each integer represents the maximum number of genes between consecutive, and `<` and `>` indicate the strand in which to search for the HMM pattern, antisense and sense, respectively. Alternative HMMs can be used for a single gene, as shown in the HMM group $(HMM_{B1} | HMM_{B2})$ above, in which case the search will be performed with both HMMs. Additionally, gene symbols can be used directly in the query string when the [@pgap] HMM database is employed, which is the default database used by `Pynteny`.
+where $HMM_a$ represents the name of the HMM modeling gene A, each integer represents the maximum number of genes between consecutive HMMs, and `<` and `>` indicate the strand in which to search for the HMM pattern, antisense and sense, respectively. Alternative HMMs can be used for a single gene, as shown in the HMM group $(HMM_{B1} | HMM_{B2})$ above, in which case the search will be performed with both HMMs. Additionally, gene symbols can be used directly in the query string when the [@pgap] HMM database is employed, which is the default database used by `Pynteny`.
 
 `Pynteny` was designed to be used by researchers working with large, unannotated sequence databases, such as those typically encountered in metagenomic analyses. It can be accessed through a command line interface or easily integrated into pipelines as a Python package. It can directly handle assembled nucleotide sequence data, however, it also accepts annotated genomes in GenBank format as input data. In both cases, the package provides all the necessary functionality to preprocess the sequence database. 
 
