+++ .github/workflows/docker.yml
@@ -12,7 +12,6 @@ name: Publish Docker image
 on:
   release:
     types: [published]
-  push:
 
 jobs:
   push_to_registries:
