### 2023-03-22 #67 @Robaina
Hi @arianesasso,

just to inform you that Pynteny has been published in [JOSS](https://github.com/openjournals/joss-reviews/issues/5289).

Thank you!
